+++ src/input-parser/CSVParser.js
@@ -7,8 +7,7 @@ class CsvParser {
     const string = helper.readFileCSV(inputPath, options);
 
     const o = {
-      delimiter: ',',
-      quote: '"',
+      delimiter: (options.csv && options.csv.delimiter) ? options.csv.delimiter : ',',
     };
 
     const result = csv.toObject(string, o);